views folder, move back licenses
