+++ src/www/app/app.ts
@@ -15,8 +15,8 @@
 import * as errors from "../model/errors";
 import * as events from "../model/events";
 import {Server} from "../model/server";
-import {ServerConnectionState} from "../ui_components/servers_view/server_connection_indicator";
-import {ServerListItem} from "../ui_components/servers_view/server_list";
+import {ServerConnectionState} from "../views/servers_view/server_connection_indicator";
+import {ServerListItem} from "../views/servers_view/server_list";
 
 import {Clipboard} from "./clipboard";
 import {EnvironmentVariables} from "./environment";
