+++ src/www/ui_components/app-root.js
@@ -44,7 +44,7 @@ import "./language-view.js";
 import "./licenses-view.js";
 import "./outline-icons.js";
 import "./privacy-view.js";
-import "./servers_view";
+import "../views/servers_view";
 import "./server-rename-dialog.js";
 import "./user-comms-dialog.js";
