+++ src/www/ui_components/licenses/licenses.txt
