+++ src/www/views/servers_view/server_card/index.ts
